requirements.txt
chardet==4.0.0
numpy==1.20.1
pandas==1.2.3
pdf2image==1.10.0
pdfminer.six==20200517
pdfplumber==0.5.27
Pillow==8.1.2
pycryptodome==3.10.1
PyPDF2==1.26.0
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sortedcontainers==2.3.0
Wand==0.6.6
